Renaming `WUBRG` to `wubrg`, and refining imports.
