+++ game_metadata/GameMetadata.py
@@ -4,7 +4,7 @@ from functools import cmp_to_key
 from datetime import date, time, datetime, timedelta
 
 from Utilities.auto_logging import logging
-from WUBRG import index_dist_wubrg
+from wubrg import index_dist_wubrg
 
 from game_metadata.utils.settings import SET_CONFIG
 from game_metadata.CallScryfall import CallScryfall
